Install gcloud in khan-dotfiles
This finishes some of the last few manual steps in dotfiles setup:
installing and authing `gcloud`, and downloading a `current.sqlite`.
It's mostly pretty straightfoward: we download and extract `gcloud` (I
put it in `~/khan/devtools` rather than `~` as it defaults to), do the
login (which requires user input), then download `current.sqlite`,
skipping any steps that are already done.

I opted to just download the archive so we get a consistent experience
across all platforms.  On linux, we could do `curl | bash`, but that
gets us an interactive installer, which is not what we want.  On mac, we
could alternately do `brew install google-cloud-sdk`, but I figured it
was easier to just be consistent.  Plus, this means we can add gcloud
autocomplete to bash easily, because we know where it lives.

Test plan:
On both linux and mac:
- (comment out other parts of setup.sh, to make things faster)
- make sure the new gcloud path is not yet in my dotfiles
- if I already have gcloud installed (to a different location),
  `./setup.sh` should be a no-op
- remove any `gcloud` I have installed, and move my `current.sqlite`;
  then `./setup.sh` should install `gcloud`, log me in, and download a
  new one
- add the new gcloud path to my dotfiles, I should now be able to use it
  as normal
- `./setup.sh` should now be a no-op again
- `gcloud auth revoke`, then `./setup.sh` should log me back in
- `./setup.sh` should now be a no-op again, and gcloud should still work