Add script that does some basic analysis of develoment environment
Summary:
The ./system_report.sh script will produce a report of some of various tool versions, locations, and other local problems that have been problematic in dev-support calls in the past.

It is intended to be used early in the process of responding to a dev-support call, and uploaded via slack or other means to the dev-supporter so that they can scan the info for things that look incorrect.

Issue: INFRA-5429

Test Plan:
Run script on Mac
Run script on Linux
Intentionally break some stuff, then run the script again

Subscribers: csilvers

Differential Revision: https://phabricator.khanacademy.org/D66625